backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
async==0.6.2
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.2.0
bcrypt==4.2.1
billiard==3.6.4.0
blinker==1.9.0
boto3==1.26.68
botocore==1.29.68
cachetools==5.5.0
celery==5.2.7
certifi==2024.8.30
charset-normalizer==2.1.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Deprecated==1.2.15
Flask==2.2.2
Flask-Cors==3.0.10
frozenlist==1.5.0
google-api-core==2.10.2
google-auth==2.13.0
google-cloud-vision==3.0.0
googleapis-common-protos==1.63.1
grpcio==1.68.1
grpcio-status==1.48.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kombu==5.4.2
MarkupSafe==3.0.2
multidict==6.1.0
mysql-connector-python==8.0.30
numpy==2.2.0
openai==0.27.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==3.19.6
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyMuPDF==1.25.1
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
pytz==2024.2
redis==4.3.4
requests==2.28.1
rsa==4.9
s3transfer==0.6.2
setuptools==75.6.0
six==1.17.0
tqdm==4.67.1
tzdata==2024.2
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
Werkzeug==2.2.3
wheel==0.45.1
wrapt==1.17.0
yarl==1.18.3


# Database
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# API and XML Processing
requests==2.31.0
lxml==4.9.3

# Data Processing
pandas==2.1.1
numpy==1.26.1

# GUI
tk==0.1.0
tkcalendar==1.6.1

# Utilities
logging==0.4.9.6